refactor: Models 파일명 통일 및 Barrel 패턴 적용
