+++ src/models/Channel.ts
