+++ src/models/index.ts
@@ -1,0 +1,5 @@
+export * from './Channel'
+export * from './MatchHistory'
+export * from './RankStat'
+export * from './Schedule'
+export * from './Summoner'
